RAD-234 Move DicomUtils to package dicom
